Support for @sorin-ionescu rewrite of oh-my-zsh
Sorin's OMZ rewrite is a bit nicer, as it's written in more
native style ZSH, doesn't do auto updates, and incorporates
many bugfixes, and is more actively maintained than Robby's branch.

For now, YADR supports both, so if you want to try out the sorin
branch, go to sorin-ionescu/oh-my-zsh, install it, and rerun
the YADR installer so that the skwp theme is installed for you.

If you want to have them side by side for comparison, rename the
old robby one to ~/.oh-my-zsh.robby or similar, so you can toggle
back and forth through renames, or symlinks.

+++ Rakefile
@@ -52,7 +52,13 @@ task :install => [:submodules] do
 end
 
 task :zsh_themes do
-  run %{ ln -nfs #{ENV["PWD"]}/oh-my-zsh/themes/* $HOME/.oh-my-zsh/themes/ } if want_to_install?('zsh themes')
+  if File.exist?("#{ENV['HOME']}/.oh-my-zsh/modules/prompt/functions")
+    puts "Detected oh-my-zsh @sorin-ionescu version."
+    run %{ ln -nfs #{ENV["PWD"]}/oh-my-zsh/modules/prompt/functions/* $HOME/.oh-my-zsh/modules/prompt/functions/ } if want_to_install?('zsh themes')
+  elsif File.exist("#{ENV['HOME']}/.oh-my-zsh")
+    puts "Detected oh-my-zsh @robbyrussell version."
+    run %{ ln -nfs #{ENV["PWD"]}/oh-my-zsh/themes/* $HOME/.oh-my-zsh/themes/ } if want_to_install?('zsh themes')
+  end
 end
 
 desc "Init and update submodules."
